Fix GL context initialization on Linux NVidia drivers
