+++ src/graphics/opengl.rs
@@ -4,10 +4,8 @@ use std::ptr;
 
 use gl::{self, types::*};
 use glm::Mat4;
-use sdl2::{
-    video::{GLContext, GLProfile, Window},
-    VideoSubsystem,
-};
+use sdl2::video::{GLContext, Window};
+use sdl2::VideoSubsystem;
 
 use crate::error::{Result, TetraError};
 
@@ -24,17 +22,6 @@ pub struct GLDevice {
 
 impl GLDevice {
     pub fn new(video: &VideoSubsystem, window: &Window, vsync: bool) -> Result<GLDevice> {
-        let gl_attr = video.gl_attr();
-
-        gl_attr.set_context_profile(GLProfile::Core);
-        gl_attr.set_context_version(3, 2);
-        gl_attr.set_red_size(8);
-        gl_attr.set_green_size(8);
-        gl_attr.set_blue_size(8);
-        gl_attr.set_alpha_size(8);
-        gl_attr.set_double_buffer(true);
-        // TODO: Will need to add some more here if we start using the depth/stencil buffers
-
         let _ctx = window.gl_create_context().map_err(TetraError::OpenGl)?;
         gl::load_with(|name| video.gl_get_proc_address(name) as *const _);
